requirements.txt
simplejson
django
Pillow
torch
pytorch-ignite
transformers==2.5.1
tensorboardX==1.8
tensorflow  # for tensorboardX
djangorestframework